refactor(Statistics): 統計サービスを単一責任の原則に基づき再構築
- StatisticsServiceをSRPに従って分割
- 月別統計/変化率計算/フィルタリング/集計を独立モジュール化
- 旧実装（calculateWorkoutStats）を削除し新実装に統一
- 各モジュールが単一責任を持つ構造に変更
